+++ core/commands/add/add.go
@@ -13,8 +13,8 @@ import (
 	ccutil "github.com/jbenet/go-ipfs/core/commands/util"
 	pinning "github.com/jbenet/go-ipfs/core/corerepo/pin"
 	coreunix "github.com/jbenet/go-ipfs/core/coreunix"
-	importer "github.com/jbenet/go-ipfs/importer"
-	"github.com/jbenet/go-ipfs/importer/chunk"
+	importer "github.com/jbenet/go-ipfs/unixfs/importer"
+	"github.com/jbenet/go-ipfs/unixfs/importer/chunk"
 	dag "github.com/jbenet/go-ipfs/struct/merkledag"
 	ft "github.com/jbenet/go-ipfs/unixfs"
 	u "github.com/jbenet/go-ipfs/util"
