+++ core/coreunix/add.go
@@ -10,8 +10,8 @@ import (
 	"github.com/jbenet/go-ipfs/thirdparty/commands/files"
 	core "github.com/jbenet/go-ipfs/core"
 	"github.com/jbenet/go-ipfs/core/corerepo/pin"
-	importer "github.com/jbenet/go-ipfs/importer"
-	chunk "github.com/jbenet/go-ipfs/importer/chunk"
+	importer "github.com/jbenet/go-ipfs/unixfs/importer"
+	chunk "github.com/jbenet/go-ipfs/unixfs/importer/chunk"
 	merkledag "github.com/jbenet/go-ipfs/struct/merkledag"
 	"github.com/jbenet/go-ipfs/thirdparty/eventlog"
 	unixfs "github.com/jbenet/go-ipfs/unixfs"
