+++ struct/merkledag/merkledag_test.go
@@ -13,8 +13,8 @@ import (
 	dssync "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore/sync"
 	"github.com/jbenet/go-ipfs/core/corerepo/pin"
 	offline "github.com/jbenet/go-ipfs/exchange/offline"
-	imp "github.com/jbenet/go-ipfs/importer"
-	chunk "github.com/jbenet/go-ipfs/importer/chunk"
+	imp "github.com/jbenet/go-ipfs/unixfs/importer"
+	chunk "github.com/jbenet/go-ipfs/unixfs/importer/chunk"
 	bserv "github.com/jbenet/go-ipfs/struct/blocks/blockservice"
 	bstore "github.com/jbenet/go-ipfs/struct/blocks/blockstore"
 	. "github.com/jbenet/go-ipfs/struct/merkledag"
