+++ unixfs/fuse/ipns/ipns_unix.go
@@ -16,7 +16,7 @@ import (
 	eventlog "github.com/jbenet/go-ipfs/thirdparty/eventlog"
 
 	core "github.com/jbenet/go-ipfs/core"
-	chunk "github.com/jbenet/go-ipfs/importer/chunk"
+	chunk "github.com/jbenet/go-ipfs/unixfs/importer/chunk"
 	nsys "github.com/jbenet/go-ipfs/namesys"
 	ci "github.com/jbenet/go-ipfs/p2p/crypto"
 	path "github.com/jbenet/go-ipfs/path"
