+++ unixfs/io/dagmodifier_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 	"github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore/sync"
 	"github.com/jbenet/go-ipfs/exchange/offline"
-	imp "github.com/jbenet/go-ipfs/importer"
-	"github.com/jbenet/go-ipfs/importer/chunk"
+	imp "github.com/jbenet/go-ipfs/unixfs/importer"
+	"github.com/jbenet/go-ipfs/unixfs/importer/chunk"
 	bs "github.com/jbenet/go-ipfs/struct/blocks/blockservice"
 	"github.com/jbenet/go-ipfs/struct/blocks/blockstore"
 	mdag "github.com/jbenet/go-ipfs/struct/merkledag"
